requirements.txt
Flask==3.0.0
Pillow
pytesseract
opencv-python-headless
pandas
numpy
pywebview
pyinstaller
python-dotenv
